feat: 계산기 input 부분완료

+++ exchange_rate_calculator/exchange_rate_calculator/View/ExchangeTextViewModel.swift
@@ -8,9 +8,9 @@
 import Foundation
 
 class ExchangeTextViewModel: ObservableObject {
-    @Published var viewModelValue = 0
+    @Published var viewModelValue = 0.0
 
-    func chagne(_ value: Int) {
+    func chagne(_ value: Double) {
         viewModelValue = 10 * value
     }
 }
